Use <tar> instead of <rom> in VFS
The ROM plugin is inherently incompatible with libziparchives
handling of files, as it compares the file size with header
information. As ROMs are rounded up to full pages (which is
reflected by the announced file size), this triggeres an assertion
when opening the zip file.

This fixes the issue in ProfileCompilationInfoTest.SaveArtMethods.
Ref. #100